requirements.txt
Django==4.2.10
django-cors-headers==4.3.1
django-redis==5.4.0
django-filter==23.5
djangorestframework==3.14.0
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
faiss-cpu==1.7.4
numpy==1.26.2
redis==5.0.1
Pillow==10.1.0
markdown==3.5.1
anthropic==0.7.8